Move PlanUtils.kt and PlansConstants.java to util package
These utility classes are used across multiple features (domains, themes,
media, settings, plugins) and are not UI components, so they belong in
the util package rather than ui/plans. This also allows the now-empty
ui/plans directory to be removed.
