+++ WordPress/src/main/java/org/wordpress/android/util/PlansConstants.java
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.plans;
+package org.wordpress.android.util;
 
 
 public class PlansConstants {
